update start sim request

+++ PersonaBackend/Controllers/HandOfZeusController.cs
@@ -41,23 +41,23 @@ namespace PersonaBackend.Controllers
         {
             try
             {
-                await DeleteAllDataAsync();
+                if (request.action == "reset")
+                {
+                    await DeleteAllDataAsync();
+                    return Ok(new ApiResponse<bool> { Success = true, Data = true, Message = $"Simulation reset successfully" });
+                }
 
                 //TODO check  request data VALIDATE
-                _chronos.SetSimulationStartDate(request.StartDate);
+                _chronos.SetSimulationStartDate(request.startTime);
+                var numberOfPersonas = 1000;
 
                 await _awsManagerService.PutParameterAsync("/simulation/date", _chronos.GetCurrentDateString());
-
-                if (request.NumberOfPersonas < 1 || request.NumberOfPersonas > 50000)
-                {
-                    return BadRequest(new ApiResponse<bool> { Data = false, Message = "Invalid request. The number of personas must be between 1 and 50,000." });
-                }
 
                 var personas = new List<Persona>();
                 var events = new List<EventOccurred>();
                 var personaIDs = new List<long>();
 
-                for (int i = 0; i < request.NumberOfPersonas; i++)
+                for (int i = 0; i < numberOfPersonas; i++)
                 {
                     var persona = new Persona
                     {
@@ -118,7 +118,7 @@ namespace PersonaBackend.Controllers
                 await _dbContext.SaveChangesAsync();
 
                 await _awsManagerService.EnableSchedule("sim-schedule", true);
-                return Ok(new ApiResponse<bool> { Success = true, Data = true, Message = $"Simulation started successfully with {request.NumberOfPersonas} persona records" });
+                return Ok(new ApiResponse<bool> { Success = true, Data = true, Message = $"Simulation started successfully with {numberOfPersonas} persona records" });
             }
             catch (Exception ex)
             {
